+++ Hurtownia/Windows/AddProductWindow.xaml.cs
@@ -31,9 +31,9 @@ namespace Hurtownia.Windows
                 var name = TextBoxName.Text;
                 var code = TextBoxCode.Text;
                 var ean = TextBoxEan.Text;
-                var price = float.Parse(TextBoxPrice.Text);
+                var price = double.Parse(TextBoxPrice.Text);
                 var unit = (Product.Unit) Enum.Parse(typeof(Product.Unit), ComboBoxUnit.Text);
-                var quantity = float.Parse(TextBoxQuantity.Text);
+                var quantity = double.Parse(TextBoxQuantity.Text);
                 var newProduct = new Product(name, code, ean, price, quantity, unit);
                 Products.AddProduct(newProduct);
                 Close();
